requirements.txt
asn1crypto==0.24.0
attrs==20.2.0
Automat==0.7.0
cffi==1.11.5
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
lxml==4.2.5
parsel==1.5.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pymongo==3.7.2
pyOpenSSL==18.0.0
queuelib==1.5.0
redis==3.0.1
Scrapy==1.5.1
scrapy-redis==0.6.8
service-identity==17.0.0
six==1.11.0
Twisted==20.3.0
w3lib==1.19.0
zope.interface==4.6.0
PyExecJS==1.5.1
selenium==3.141.0
urllib3==1.25.10
scrapyd==1.2.1
scrapyd-client==1.1.0
